문서 정보 표시 제작 25 - 로딩 페이지 분리
