+++ app/words-docs/[id]/DocsDataHome.tsx
@@ -5,9 +5,7 @@ import NotFound from "@/app/not-found-client";
 import ErrorPage from "@/app/components/ErrorPage";
 import { useState, useCallback, useEffect } from "react";
 import type { PostgrestError } from "@supabase/supabase-js";
-import ProgressBar from "@/app/components/ProgressBar";
-import type { LoadingState } from "@/app/types/type";
-import Spinner from "@/app/components/Spinner";
+import LoadingPage, {useLoadingState } from '@/app/components/LoadingPage';
 
 type wordsDataType = ({
     word: string;
@@ -25,21 +23,9 @@ type wordsDataType = ({
 
 export default function DocsDataHome({id}:{id:number}){
     const [isNotFound,setIsNotFound] = useState(false);
-    const [loadingState, setLoadingState] = useState<LoadingState>({
-                isLoading: true,
-                progress: 0,
-                currentTask: "초기화 중..."
-            });
+    const { loadingState, updateLoadingState } = useLoadingState();
     const [errorMessage,setErrorMessage] = useState<string|null>(null);
     const [wordsData,setWordsData] = useState<{words:wordsDataType[], metadata:{title:string, lastUpdate:string}} | null>(null);
-
-    const updateLoadingState = (progress: number, task: string) => {
-        setLoadingState({
-            isLoading: progress < 100,
-            progress,
-            currentTask: task
-        });
-    };
 
     const getDataOkWords = useCallback(async () => {
         const {data:dataA, error:error} = await supabase.from('docs_words').select('words(word)').eq('docs_id',id);
@@ -141,19 +127,7 @@ export default function DocsDataHome({id}:{id:number}){
 
     if (loadingState.isLoading) {
         return (
-            <div className="flex flex-col items-center justify-center p-8 bg-white rounded-lg shadow min-h-screen min-w-full">
-                <h2 className="text-xl font-bold mb-4">단어 정보 로딩 중</h2>
-                <div className="w-full max-w-md mb-4">
-                    <ProgressBar
-                        completed={loadingState.progress}
-                        label={`${loadingState.progress}% 완료`}
-                    />
-                </div>
-                <p className="text-gray-600 mt-2">{loadingState.currentTask}</p>
-                <div className="mt-4">
-                    <Spinner />
-                </div>
-            </div>
+            <LoadingPage title={"문서"} />
         );
     }
