+++ app/words-docs/[id]/info/DocsInfoPage.tsx
@@ -3,10 +3,8 @@ import { supabase } from "@/app/lib/supabaseClient";
 import NotFound from "@/app/not-found-client";
 import ErrorPage from "@/app/components/ErrorPage"
 import { useCallback, useEffect, useState } from "react";
-import type { LoadingState } from "@/app/types/type";
-import Spinner from "@/app/components/Spinner";
-import ProgressBar from "@/app/components/ProgressBar";
 import DocsInfo from "./DocsInfo";
+import LoadingPage, {useLoadingState } from '@/app/components/LoadingPage';
 
 type docsInfoType = {
     metadata: {
@@ -24,21 +22,9 @@ type docsInfoType = {
 
 export default function DocsInfoPage({ id }: { id: number }) {
     const [isNotFound, setIsNotFound] = useState(false);
-    const [loadingState, setLoadingState] = useState<LoadingState>({
-        isLoading: true,
-        progress: 0,
-        currentTask: "초기화 중..."
-    });
     const [errorMessage, setErrorMessage] = useState<string | null>(null);
     const [docsInfoData, setDocsInfoData] = useState<docsInfoType | null>(null);
-
-    const updateLoadingState = (progress: number, task: string) => {
-        setLoadingState({
-            isLoading: progress < 100,
-            progress,
-            currentTask: task
-        });
-    };
+    const { loadingState, updateLoadingState } = useLoadingState();
 
     const getDocs = useCallback(async () => {
         const { data, error } = await supabase.from('docs').select('id, created_at, name, users(nickname), typez, last_update').eq('id', id).maybeSingle();
@@ -102,19 +88,7 @@ export default function DocsInfoPage({ id }: { id: number }) {
 
     if (loadingState.isLoading) {
         return (
-            <div className="flex flex-col items-center justify-center p-8 bg-white rounded-lg shadow min-h-screen min-w-full">
-                <h2 className="text-xl font-bold mb-4">단어 정보 로딩 중</h2>
-                <div className="w-full max-w-md mb-4">
-                    <ProgressBar
-                        completed={loadingState.progress}
-                        label={`${loadingState.progress}% 완료`}
-                    />
-                </div>
-                <p className="text-gray-600 mt-2">{loadingState.currentTask}</p>
-                <div className="mt-4">
-                    <Spinner />
-                </div>
-            </div>
+            <LoadingPage title={"문서 정보"}/>
         );
     }
