+++ app/words-docs/[id]/logs/DocsLogPage.tsx
@@ -3,10 +3,8 @@ import DocsLogs from "./DocsLogs";
 import NotFound from "@/app/not-found-client";
 import { supabase } from "@/app/lib/supabaseClient";
 import { useState, useEffect, useCallback } from 'react';
-import ProgressBar from "@/app/components/ProgressBar";
-import type { LoadingState } from "@/app/types/type";
-import Spinner from "@/app/components/Spinner";
 import ErrorPage from "@/app/components/ErrorPage";
+import LoadingPage, {useLoadingState } from '@/app/components/LoadingPage';
 
 type log = {
     id: number;
@@ -18,21 +16,9 @@ type log = {
 
 export default function DocsLogPage({id}:{id: number}){
     const [isNotFound,setIsNotFound] = useState(false);
-    const [loadingState, setLoadingState] = useState<LoadingState>({
-            isLoading: true,
-            progress: 0,
-            currentTask: "초기화 중..."
-        });
     const [errorMessage,setErrorMessage] = useState<string|null>(null);
     const [logsData, setLogsDatas] = useState<{logs:log[], docsName: string}|null>(null);
-
-    const updateLoadingState = (progress: number, task: string) => {
-        setLoadingState({
-            isLoading: progress < 100,
-            progress,
-            currentTask: task
-        });
-    };
+    const { loadingState, updateLoadingState } = useLoadingState();
 
     const getDocs = useCallback(async () => {
         const { data, error } = await supabase
@@ -92,19 +78,7 @@ export default function DocsLogPage({id}:{id: number}){
     
     if (loadingState.isLoading) {
         return (
-            <div className="flex flex-col items-center justify-center p-8 bg-white rounded-lg shadow min-h-screen min-w-full">
-                <h2 className="text-xl font-bold mb-4">단어 정보 로딩 중</h2>
-                <div className="w-full max-w-md mb-4">
-                    <ProgressBar
-                        completed={loadingState.progress}
-                        label={`${loadingState.progress}% 완료`}
-                    />
-                </div>
-                <p className="text-gray-600 mt-2">{loadingState.currentTask}</p>
-                <div className="mt-4">
-                    <Spinner />
-                </div>
-            </div>
+            <LoadingPage title={"문서 로그"} />
         );
     }
